examples/requirements.txt
firecrawl-py
python-dotenv
openai
